make config available in app state
