+++ Cargo.toml
@@ -9,7 +9,6 @@ axum = "0.6"
 axum-on-rails = { path = "./axum-on-rails" }
 clap = "4.4"
 dotenvy = "0.15"
-figment = { version = "0.10", features = ["toml"] }
 rand = "0.8"
 refinery = { version = "0.8", features = ["tokio-postgres"]}
 serde = { version = "1.0", features = ["derive"] }
