+++ axum-on-rails/src/lib.rs
@@ -1,2 +1,8 @@
 pub mod cli;
 pub mod test;
+
+mod util;
+
+pub use util::get_env;
+pub use util::load_config;
+pub use util::Environment;
