+++ src/state.rs
@@ -1,12 +1,14 @@
+use crate::config::Config;
 use sqlx::postgres::{PgPool, PgPoolOptions};
 use std::env;
 
 #[derive(Clone)]
 pub struct AppState {
     pub db_pool: PgPool,
+    pub config: Config,
 }
 
-pub async fn app_state() -> AppState {
+pub async fn app_state(config: Config) -> AppState {
     let db_url = env::var("DATABASE_URL").expect("No DATABASE_URL set – cannot start server!");
 
     let db_pool = PgPoolOptions::new()
@@ -14,5 +16,5 @@ pub async fn app_state() -> AppState {
         .await
         .expect("Could not connect to database!");
 
-    AppState { db_pool }
+    AppState { db_pool, config }
 }
